DUP move volatile triggers functions from impl to core.triggers and test missing lines
