+++ apps/astarte_data_updater_plant/lib/astarte_data_updater_plant/data_updater/impl.ex
@@ -20,15 +20,8 @@ defmodule Astarte.DataUpdaterPlant.DataUpdater.Impl do
   alias Astarte.DataUpdaterPlant.DataUpdater.Core
   alias Astarte.DataUpdaterPlant.Config
   alias Astarte.Core.Device
-  alias Astarte.Core.InterfaceDescriptor
-  alias Astarte.Core.Mapping.EndpointsAutomaton
   alias Astarte.DataUpdaterPlant.DataUpdater.State
-  alias Astarte.Core.Triggers.DataTrigger
-  alias Astarte.Core.Triggers.SimpleTriggersProtobuf.DataTrigger, as: ProtobufDataTrigger
-  alias Astarte.Core.Triggers.SimpleTriggersProtobuf.Utils, as: SimpleTriggersProtobufUtils
-  alias Astarte.DataAccess.Interface, as: InterfaceQueries
   alias Astarte.DataUpdaterPlant.DataUpdater.Cache
-  alias Astarte.DataUpdaterPlant.DataUpdater.EventTypeUtils
   alias Astarte.DataUpdaterPlant.DataUpdater.Queries
   alias Astarte.DataUpdaterPlant.MessageTracker
   alias Astarte.DataUpdaterPlant.TriggersHandler
@@ -179,219 +172,4 @@ defmodule Astarte.DataUpdaterPlant.DataUpdater.Impl do
 
   defdelegate handle_introspection(state, payload, message_id, timestamp),
     to: Core.IntrospectionHandler
-
-  def handle_install_volatile_trigger(
-        %State{discard_messages: true} = state,
-        _,
-        message_id,
-        _
-      ) do
-    MessageTracker.ack_delivery(state.message_tracker, message_id)
-    state
-  end
-
-  def handle_install_volatile_trigger(
-        state,
-        object_id,
-        object_type,
-        parent_id,
-        trigger_id,
-        simple_trigger,
-        trigger_target
-      ) do
-    trigger = SimpleTriggersProtobufUtils.deserialize_simple_trigger(simple_trigger)
-
-    target =
-      SimpleTriggersProtobufUtils.deserialize_trigger_target(trigger_target)
-      |> Map.put(:simple_trigger_id, trigger_id)
-      |> Map.put(:parent_trigger_id, parent_id)
-
-    volatile_triggers_list = [
-      {{object_id, object_type}, {trigger, target}} | state.volatile_triggers
-    ]
-
-    new_state = Map.put(state, :volatile_triggers, volatile_triggers_list)
-
-    if Map.has_key?(new_state.interface_ids_to_name, object_id) do
-      interface_name = Map.get(new_state.interface_ids_to_name, object_id)
-      %InterfaceDescriptor{automaton: automaton} = new_state.interfaces[interface_name]
-
-      case trigger do
-        {:data_trigger, %ProtobufDataTrigger{match_path: "/*"}} ->
-          {:ok, Core.Trigger.load_trigger(new_state, trigger, target)}
-
-        {:data_trigger, %ProtobufDataTrigger{match_path: match_path}} ->
-          with {:ok, _endpoint_id} <- EndpointsAutomaton.resolve_path(match_path, automaton) do
-            {:ok, Core.Trigger.load_trigger(new_state, trigger, target)}
-          else
-            {:guessed, _} ->
-              # State rollback here
-              {{:error, :invalid_match_path}, state}
-
-            {:error, :not_found} ->
-              # State rollback here
-              {{:error, :invalid_match_path}, state}
-          end
-      end
-    else
-      case trigger do
-        {:data_trigger, %ProtobufDataTrigger{interface_name: "*"}} ->
-          {:ok, Core.Trigger.load_trigger(new_state, trigger, target)}
-
-        {:data_trigger,
-         %ProtobufDataTrigger{
-           interface_name: interface_name,
-           interface_major: major,
-           match_path: "/*"
-         }} ->
-          with :ok <-
-                 InterfaceQueries.check_if_interface_exists(state.realm, interface_name, major) do
-            {:ok, new_state}
-          else
-            {:error, reason} ->
-              # State rollback here
-              {{:error, reason}, state}
-          end
-
-        {:data_trigger,
-         %ProtobufDataTrigger{
-           interface_name: interface_name,
-           interface_major: major,
-           match_path: match_path
-         }} ->
-          with {:ok, %InterfaceDescriptor{automaton: automaton}} <-
-                 InterfaceQueries.fetch_interface_descriptor(state.realm, interface_name, major),
-               {:ok, _endpoint_id} <- EndpointsAutomaton.resolve_path(match_path, automaton) do
-            {:ok, new_state}
-          else
-            {:error, :not_found} ->
-              {{:error, :invalid_match_path}, state}
-
-            {:guessed, _} ->
-              {{:error, :invalid_match_path}, state}
-
-            {:error, reason} ->
-              # State rollback here
-              {{:error, reason}, state}
-          end
-
-        {:device_trigger, _} ->
-          {:ok, Core.Trigger.load_trigger(new_state, trigger, target)}
-      end
-    end
-  end
-
-  def handle_delete_volatile_trigger(%State{discard_messages: true} = state, _, message_id, _) do
-    MessageTracker.discard(state.message_tracker, message_id)
-    state
-  end
-
-  def handle_delete_volatile_trigger(state, trigger_id) do
-    {new_volatile, maybe_trigger} =
-      Enum.reduce(state.volatile_triggers, {[], nil}, fn item, {acc, found} ->
-        {_, {_simple_trigger, trigger_target}} = item
-
-        if trigger_target.simple_trigger_id == trigger_id do
-          {acc, item}
-        else
-          {[item | acc], found}
-        end
-      end)
-
-    case maybe_trigger do
-      {{obj_id, obj_type}, {simple_trigger, trigger_target}} ->
-        %{state | volatile_triggers: new_volatile}
-        |> delete_volatile_trigger({obj_id, obj_type}, {simple_trigger, trigger_target})
-
-      nil ->
-        {:ok, state}
-    end
-  end
-
-  defp delete_volatile_trigger(
-         state,
-         {obj_id, _obj_type},
-         {{:data_trigger, proto_buf_data_trigger}, trigger_target_to_be_deleted}
-       ) do
-    if Map.get(state.interface_ids_to_name, obj_id) do
-      data_trigger_to_be_deleted =
-        SimpleTriggersProtobufUtils.simple_trigger_to_data_trigger(proto_buf_data_trigger)
-
-      data_triggers = state.data_triggers
-
-      event_type =
-        EventTypeUtils.pretty_data_trigger_type(proto_buf_data_trigger.data_trigger_type)
-
-      data_trigger_key =
-        Core.DataTrigger.data_trigger_to_key(state, data_trigger_to_be_deleted, event_type)
-
-      existing_triggers_for_key = Map.get(data_triggers, data_trigger_key, [])
-
-      # Separate triggers for key between the trigger congruent with the one being deleted
-      # and all the other triggers
-      {congruent_data_trigger_for_key, other_data_triggers_for_key} =
-        Enum.reduce(existing_triggers_for_key, {nil, []}, fn
-          trigger, {congruent_data_trigger_for_key, other_data_triggers_for_key} ->
-            if DataTrigger.are_congruent?(trigger, data_trigger_to_be_deleted) do
-              {trigger, other_data_triggers_for_key}
-            else
-              {congruent_data_trigger_for_key, [trigger | other_data_triggers_for_key]}
-            end
-        end)
-
-      next_data_triggers_for_key =
-        case congruent_data_trigger_for_key do
-          nil ->
-            # Trying to delete an unexisting volatile trigger, just return old data triggers
-            existing_triggers_for_key
-
-          %DataTrigger{trigger_targets: [^trigger_target_to_be_deleted]} ->
-            # The target of the deleted trigger was the only target, just remove it
-            other_data_triggers_for_key
-
-          %DataTrigger{trigger_targets: targets} ->
-            # The trigger has other targets, drop the one that is being deleted and update
-            new_trigger_targets = Enum.reject(targets, &(&1 == trigger_target_to_be_deleted))
-
-            new_congruent_data_trigger_for_key = %{
-              congruent_data_trigger_for_key
-              | trigger_targets: new_trigger_targets
-            }
-
-            [new_congruent_data_trigger_for_key | other_data_triggers_for_key]
-        end
-
-      next_data_triggers =
-        if is_list(next_data_triggers_for_key) and length(next_data_triggers_for_key) > 0 do
-          Map.put(data_triggers, data_trigger_key, next_data_triggers_for_key)
-        else
-          Map.delete(data_triggers, data_trigger_key)
-        end
-
-      {:ok, %{state | data_triggers: next_data_triggers}}
-    else
-      {:ok, state}
-    end
-  end
-
-  defp delete_volatile_trigger(
-         state,
-         {_obj_id, _obj_type},
-         {{:device_trigger, proto_buf_device_trigger}, trigger_target}
-       ) do
-    event_type =
-      EventTypeUtils.pretty_device_event_type(proto_buf_device_trigger.device_event_type)
-
-    device_triggers = state.device_triggers
-
-    updated_targets_list =
-      Map.get(device_triggers, event_type, [])
-      |> Enum.reject(fn target ->
-        target == trigger_target
-      end)
-
-    updated_device_triggers = Map.put(device_triggers, event_type, updated_targets_list)
-
-    {:ok, %{state | device_triggers: updated_device_triggers}}
-  end
 end
